tpu_requirements.txt
accelerate
chex
typing
fjformer
transformers
einops
optax
msgpack
ipython
tqdm
pydantic
datasets
setuptools
gradio
termcolor
wandb
tensorboard
numpy
uvicorn
fastapi
pydantic-core
gcsfs
fire
google-cloud-storage